+++ src/main/java/com/brumethon/app/infrastructure/service/UserService.java
@@ -5,7 +5,7 @@ import com.brumethon.app.domain.user.UserRepository;
 import com.brumethon.app.domain.user.UserValidator;
 import com.brumethon.kernel.SimpleService;
 
-public class UserService extends SimpleService<UserRepository, User, Integer> {
+public class UserService extends SimpleService<UserRepository, User, Long> {
 
     public UserService(UserRepository repository, UserValidator validator) {
         super(repository, validator, "user");
